Former-commit-id: ca99dc714c33c05cb30e0851ca340aca5b323fea

+++ Libs/libunibreak/src/wordbreak.c
@@ -81,7 +81,7 @@ static enum WordBreakClass get_char_wb_class(
         size_t len)
 {
     int min = 0;
-    int max = len - 1;
+    int max = (int)(len - 1);
     int mid;
 
     do
